+++ NEWS.md
@@ -1,3 +1,10 @@
+# swirlify 0.4
+
+* Removed `swirlify_help()`.
+
+* Changed `demo_lesson()` to `testit()`, although it will be changed back before
+swirlify 0.5. This was done because of compatibility issues with swirl.
+
 # swirlify 0.3.3
 
 * `test_lesson()` and `test_course()` now check for proper lesson formatting.